fix (?) parallel unit forward

+++ exp/contributions.py
@@ -28,7 +28,7 @@ def main(
     checkpoint_idx: int | None = None,
     maxlen: int = 512,
     device: str = "cuda",
-    dtype: str = "bf16",
+    dtype: str = "fp16",
     load_in_8bit: bool = False,
     load_in_4bit: bool = False,
     n: int = 3,
