+++ exp/contributions_sweep.py
@@ -14,7 +14,7 @@ def sweep(
     dataset_name: str = "redpajama-1",
     maxlen: int = 128,
     device: str = "cuda",
-    dtype: str = "bf16",
+    dtype: str = "fp16",
     load_in_8bit: bool = False,
     load_in_4bit: bool = False,
     n: int = 3,
